// src/server/angelScriptTokenTables.generated.ts
// This file is auto-generated by scripts/conformance/syncAngelScriptTokenDefs.mjs.
// Do not edit manually.
// Source: https://raw.githubusercontent.com/anjo76/angelscript/master/sdk/angelscript/source/as_tokendef.h
// Generated at: 2026-02-22T00:36:26.525Z

export const angelScriptReservedKeywords = [
  "and",
  "auto",
  "bool",
  "break",
  "case",
  "cast",
  "catch",
  "class",
  "const",
  "continue",
  "default",
  "do",
  "double",
  "else",
  "enum",
  "false",
  "float",
  "for",
  "foreach",
  "funcdef",
  "if",
  "import",
  "in",
  "inout",
  "int",
  "int8",
  "int16",
  "int32",
  "int64",
  "interface",
  "is",
  "mixin",
  "namespace",
  "not",
  "null",
  "or",
  "out",
  "private",
  "protected",
  "return",
  "switch",
  "true",
  "try",
  "typedef",
  "uint",
  "uint8",
  "uint16",
  "uint32",
  "uint64",
  "using",
  "void",
  "while",
  "xor",
] as const;

export const angelScriptContextualKeywords = [
  "this",
  "from",
  "super",
  "shared",
  "final",
  "override",
  "get",
  "set",
  "abstract",
  "function",
  "if_handle_then_const",
  "external",
  "explicit",
  "property",
  "delete",
] as const;

export const angelScriptKeywordLikeTokens = [
  ...angelScriptReservedKeywords,
  ...angelScriptContextualKeywords,
] as const;